move event getter into prismic service
