+++ server/controllers/content.js
@@ -12,8 +12,7 @@ import {
   getArticleList,
   getSeriesArticles
 } from '../services/prismic-content';
-import {getArticle} from '../services/prismic';
-import {getEvent} from '../services/events';
+import {getArticle, getEvent} from '../services/prismic';
 import {getExhibition} from '../services/exhibitions';
 import {PromoListFactory} from '../model/promo-list';
 import {PaginationFactory} from '../model/pagination';
