+++ server/views/templates/event.config.js
@@ -1,5 +1,5 @@
 import {createPageConfig} from '../../model/page-config';
-import {getEvent} from '../../services/events';
+import {getEvent} from '../../services/prismic';
 import {getArticleStubs} from '../../services/wordpress';
 import {PromoFactory} from '../../model/promo';
